Finalization of refactoring: new folder organization for wave solvers
