+++ src/coreComponents/physicsSolvers/CMakeLists.txt
@@ -128,31 +128,31 @@ set( physicsSolvers_headers
      surfaceGeneration/SurfaceGeneratorFields.hpp
      surfaceGeneration/kernels/surfaceGenerationKernels.hpp
      surfaceGeneration/kernels/surfaceGenerationKernelsHelpers.hpp
-     wavePropagation/WaveSolverBase.hpp
-     wavePropagation/WaveSolverUtils.hpp
-     wavePropagation/PrecomputeSourcesAndReceiversKernel.hpp
-     wavePropagation/AcousticFields.hpp
-     wavePropagation/AcousticWaveEquationSEM.hpp
-     wavePropagation/AcousticWaveEquationSEMKernel.hpp
-     wavePropagation/AcousticTimeSchemeSEMKernel.hpp
-     wavePropagation/AcousticMatricesSEMKernel.hpp
-     wavePropagation/AcousticPMLSEMKernel.hpp
-     wavePropagation/ElasticFields.hpp
-     wavePropagation/ElasticWaveEquationSEM.hpp
-     wavePropagation/ElasticWaveEquationSEMKernel.hpp
-     wavePropagation/ElasticTimeSchemeSEMKernel.hpp
-     wavePropagation/ElasticFirstOrderWaveEquationSEM.hpp
-     wavePropagation/ElasticFirstOrderWaveEquationSEMKernel.hpp
-     wavePropagation/AcousticFirstOrderWaveEquationSEM.hpp
-     wavePropagation/AcousticFirstOrderWaveEquationSEMKernel.hpp
-     wavePropagation/ElasticMatricesSEMKernel.hpp
-     wavePropagation/AcousticVTIFields.hpp
-     wavePropagation/AcousticVTIWaveEquationSEM.hpp
-     wavePropagation/AcousticVTIWaveEquationSEMKernel.hpp
-     wavePropagation/AcoustoElasticFields.hpp
-     wavePropagation/AcoustoElasticTimeSchemeSEMKernel.hpp
-     wavePropagation/AcousticElasticWaveEquationSEM.hpp
-     wavePropagation/AcousticElasticWaveEquationSEMKernel.hpp )
+     wavePropagation/shared/WaveSolverBase.hpp
+     wavePropagation/shared/WaveSolverUtils.hpp
+     wavePropagation/shared/PrecomputeSourcesAndReceiversKernel.hpp
+     wavePropagation/acoustic/shared/AcousticFields.hpp
+     wavePropagation/acoustic/secondOrderEqn/isotropic/AcousticWaveEquationSEM.hpp
+     wavePropagation/acoustic/secondOrderEqn/isotropic/AcousticWaveEquationSEMKernel.hpp
+     wavePropagation/acoustic/shared/AcousticTimeSchemeSEMKernel.hpp
+     wavePropagation/acoustic/shared/AcousticMatricesSEMKernel.hpp
+     wavePropagation/acoustic/secondOrderEqn/isotropic/AcousticPMLSEMKernel.hpp
+     wavePropagation/elastic/shared/ElasticFields.hpp
+     wavePropagation/elastic/secondOrderEqn/isotropic/ElasticWaveEquationSEM.hpp
+     wavePropagation/elastic/secondOrderEqn/isotropic/ElasticWaveEquationSEMKernel.hpp
+     wavePropagation/elastic/shared/ElasticTimeSchemeSEMKernel.hpp
+     wavePropagation/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.hpp
+     wavePropagation/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEMKernel.hpp
+     wavePropagation/acoustic/firstOrderEqn/isotropic/AcousticFirstOrderWaveEquationSEM.hpp
+     wavePropagation/acoustic/firstOrderEqn/isotropic/AcousticFirstOrderWaveEquationSEMKernel.hpp
+     wavePropagation/elastic/shared/ElasticMatricesSEMKernel.hpp
+     wavePropagation/acoustic/secondOrderEqn/vti/AcousticVTIFields.hpp
+     wavePropagation/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEM.hpp
+     wavePropagation/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEMKernel.hpp
+     wavePropagation/acoustoelastic/secondOrderEqn/isotropic/AcoustoElasticFields.hpp
+     wavePropagation/acoustoelastic/secondOrderEqn/isotropic/AcoustoElasticTimeSchemeSEMKernel.hpp
+     wavePropagation/acoustoelastic/secondOrderEqn/isotropic/AcousticElasticWaveEquationSEM.hpp
+     wavePropagation/acoustoelastic/secondOrderEqn/isotropic/AcousticElasticWaveEquationSEMKernel.hpp )
 
 # Specify solver sources
 set( physicsSolvers_sources
@@ -209,13 +209,13 @@ set( physicsSolvers_sources
      surfaceGeneration/EmbeddedSurfacesParallelSynchronization.cpp
      surfaceGeneration/ParallelTopologyChange.cpp
      surfaceGeneration/SurfaceGenerator.cpp
-     wavePropagation/WaveSolverBase.cpp
-     wavePropagation/AcousticWaveEquationSEM.cpp
-     wavePropagation/ElasticWaveEquationSEM.cpp
-     wavePropagation/ElasticFirstOrderWaveEquationSEM.cpp
-     wavePropagation/AcousticFirstOrderWaveEquationSEM.cpp
-     wavePropagation/AcousticVTIWaveEquationSEM.cpp
-     wavePropagation/AcousticElasticWaveEquationSEM.cpp )
+     wavePropagation/shared/WaveSolverBase.cpp
+     wavePropagation/acoustic/secondOrderEqn/isotropic/AcousticWaveEquationSEM.cpp
+     wavePropagation/elastic/secondOrderEqn/isotropic/ElasticWaveEquationSEM.cpp
+     wavePropagation/elastic/firstOrderEqn/isotropic/ElasticFirstOrderWaveEquationSEM.cpp
+     wavePropagation/acoustic/firstOrderEqn/isotropic/AcousticFirstOrderWaveEquationSEM.cpp
+     wavePropagation/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEM.cpp
+     wavePropagation/acoustoelastic/secondOrderEqn/isotropic/AcousticElasticWaveEquationSEM.cpp )
 
   include( solidMechanics/kernels/SolidMechanicsKernels.cmake)
   include( multiphysics/poromechanicsKernels/PoromechanicsKernels.cmake)
