+++ src/coreComponents/physicsSolvers/wavePropagation/acoustic/secondOrderEqn/vti/AcousticVTIWaveEquationSEM.cpp
@@ -26,8 +26,8 @@
 #include "mainInterface/ProblemManager.hpp"
 #include "mesh/ElementType.hpp"
 #include "mesh/mpiCommunications/CommunicationTools.hpp"
-#include "WaveSolverUtils.hpp"
-#include "AcousticTimeSchemeSEMKernel.hpp"
+#include "physicsSolvers/wavePropagation/shared/WaveSolverUtils.hpp"
+#include "physicsSolvers/wavePropagation/acoustic/shared/AcousticTimeSchemeSEMKernel.hpp"
 
 namespace geos
 {
